+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Update Attributes package dependency to 3.1.0
+
 ## [4.3.0] - 2025-05-14
 ### Changed
 - Reset TimeScale before load scene