+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - Update Attributes package dependency to 3.1.0
+- Increase Unity minimum version to 6000.1
 
 ### Removed
 - Awaitable Coroutines package dependency