+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [5.0.0] - 2025-08-06
 ### Changed
 - Update Attributes package dependency to 3.1.0
 - Increase Unity minimum version to 6000.0
@@ -117,7 +118,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - README
 - Initial commit
 
-[Unreleased]: https://github.com/HyagoOliveira/SceneManagement/compare/4.3.0...main
+[Unreleased]: https://github.com/HyagoOliveira/SceneManagement/compare/5.0.0...main
+[5.0.0]: https://github.com/HyagoOliveira/SceneManagement/tree/4.3.0/
 [4.3.0]: https://github.com/HyagoOliveira/SceneManagement/tree/4.3.0/
 [4.2.0]: https://github.com/HyagoOliveira/SceneManagement/tree/4.2.0/
 [4.1.0]: https://github.com/HyagoOliveira/SceneManagement/tree/4.1.0/